introduce-tokenized-request-builder(move): Hand TokenizedRequest assembly over to TokenizedRequestBuilder
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
_create_tokenized_object + _resolve_embed_overrides from TokenizerManager;
paste into TokenizedRequestBuilder (drop @staticmethod, replace
``self: "TokenizedRequestBuilder"`` → plain ``self``). Rename
_create_tokenized_object → build (scope-induced; method is now public API
of new class). Caller prefix replacement: ``TokenizerManager._create_tokenized_object(self.tokenized_request_builder, ...)``
→ ``self.tokenized_request_builder.build(...)``.

Refactor chain ID: introduce-tokenized-request-builder-move
PR-Title: Extract TokenizedRequestBuilder from TokenizerManager
